requirements.txt
python-dotenv==0.15.0
tweepy==3.10.0
